requirements.txt
beautifulsoup4==4.11.1
numpy==1.24.3
Pillow==9.5.0
PyYAML==6.0
requests==2.28.2
semver==2.13.0
